iommu: arm-smmu: Add no-iommu-halt enforcement with DT property
Some IOMMUv2 implementations do NOT support _any_ form of
IOMMU halting, not just regulator-driven one, but .. any.

Add a property to enforce IOMMU HALT skip: in this condition
we will not write any HALT command to the micro mmu control
register, not by regular writel and not by secure calls.
Of course, the analogous part in the resume path will also
be skipped.

Please note then that, in the SMMU resume path, we are anyway
required to restore the secure state for the secure IOMMU to
be usable again after system suspend.