media: msm: camera_v2: isp47: Preset num_norm_clk to VFE clk count
The num_norm_clk variable is getting fed to the camera SOC-API
when cycling to (de)initialize ISP clocks.
In case we have a HVX clock (vfe stream clock) it is getting set
to a specific value but, otherwise, it was left to the previously
initialized one (which is zero!), hence clocks will not get up.

This fixes clock init for all the SoCs that don't have any HVX
clock, such as the ones using the older VFE40.