drivers: bcmdhd: apply changes to dhd_sdio from 4.4 bcmdhd
These changes are taken from Samsung's bcmdhd driver, which is used on its 4.4 devices that have bcm wifi.
The changes include:
- a new iovar for suspend
- moving codeblocks that handle sdmode and sdclk further up in the code flow

These are required to remain compatible with the 4.4 API.

Thanks to kholk@github and humberos@github for pointing out the updated bcmdhd driver from samsung and narrowing the search scope to only dhd_sdio, when it came to the components that need updating.

Change-Id: I810f6102f01bf8341d27e16af3a62678e89b5b8f