arm: DT: Tone: HACK: Remove cont-splash-memory from framebuffer
Remove the cont-splash-memory from the framebuffer nodes to
avoid IOMMU related crashes at display open from userspace.
This is	necessary after	a sane QC fix on mdss_mdp_splash_logo
appeared, as commit hash 245031160bff81703cf109e473329c3f01d77304:
previously, the	PA to VA mapping was completely	broken but now,
being it sane, we need the related IOMMU domain	to be set as
UNSECURE and correctly released	by the bootloader.

This is	a HACK and will	have the side effect of	displaying
corruption for one second prior	the Android bootanimation to
kick in.